src/cafe/cafe.service.ts
import { Injectable } from '@nestjs/common';
import { InjectRepository } from '@nestjs/typeorm';
import { BusinessLogicException, BusinessError } from 'src/shared/errors/business-errors';
import { Repository } from 'typeorm';
import { CafeEntity } from './cafe.entity';

@Injectable()
export class CafeService {
    constructor(
        @InjectRepository(CafeEntity)
        private readonly cafeRepository: Repository<CafeEntity>
    ){}

    async create(cafe: CafeEntity): Promise<CafeEntity> {
        if (cafe.precio < 0)
          throw new BusinessLogicException("The cafe with the given id has negative price", BusinessError.PRECONDITION_FAILED);
        
        return await this.cafeRepository.save(cafe);
    }
}
